+++ R/data.R
@@ -38,3 +38,19 @@
 #' }
 #' @source TODO
 "input_metadata"
+
+
+#' Multipathogen input data
+#'
+#' Linelist containing multiple pathogens (Pertussis, Enterobacter, Salmonella).
+#' This dataset is a modified version of [input_example], with the pathogen column
+#' modified for lines 3000 to 3999 (Enterobacter), and 4000 to 4999 (Salmonella)
+#'
+#' @format ## `input_example_multipathogen`
+#' A data frame with 5.000 rows and 29 columns:
+#' \describe{
+#'   \item{TO}{DO}
+#'   ...
+#' }
+#' @source TODO
+"input_example_multipathogen"