+++ R/mod_tabpanel_report.R
@@ -65,13 +65,13 @@ mod_tabpanel_report_server <- function(id,
               shiny::selectInput(NS(id, "format"), "Choose a format:",
                 choices = c("HTML", "DOCX")
               ),
-              shiny::checkboxInput(NS(id, "tables"),
-                "Include tables (stratifications)",
-                value = TRUE
-              ),
-              shiny::checkboxInput(NS(id, "interactive"),
-                "Interactive HTML",
-                value = TRUE
+              # Show tables checkbox only if DOCX is selected
+              shiny::conditionalPanel(
+                condition = sprintf("input['%s'] == 'DOCX'", NS(id, "format")),
+                shiny::checkboxInput(NS(id, "tables"),
+                  "Include signals tables for strata",
+                  value = TRUE
+                )
               ),
               shiny::downloadButton(NS(id, "downloadReport"), "Create Report")
             ),
@@ -83,11 +83,11 @@ mod_tabpanel_report_server <- function(id,
       }
     })
 
-    shiny::observeEvent(input$format, {
-      if (input$format == "HTML") {
-        shinyjs::show(id = "interactive")
+    tables <- reactive({
+      if (is.null(input$tables)) {
+        FALSE # default fallback when tables checkbox is hidden (HTML)
       } else {
-        shinyjs::hide(id = "interactive")
+        input$tables
       }
     })
 
@@ -117,8 +117,7 @@ mod_tabpanel_report_server <- function(id,
           "SignalDetectionReport.",
           switch(input$format,
             HTML = "html",
-            DOCX = "docx",
-            PDF = "pdf"
+            DOCX = "docx"
           )
         )
       },
@@ -128,13 +127,13 @@ mod_tabpanel_report_server <- function(id,
           data = filtered_data(),
           method = names(available_algorithms()[which(available_algorithms() == method())]),
           number_of_weeks = number_of_weeks(),
+          pathogens = pathogen_vars(),
           strata = strat_vars(),
-          interactive = input$interactive,
-          tables = input$tables,
+          tables = tables(),
           output_file = con,
           output_dir = NULL,
-          signals_padded = signals_padded(),
-          signals_agg = signals_agg(),
+          signals_padded = signals_padded() %>% dplyr::mutate(pathogen = pathogen_vars()),
+          signals_agg = signals_agg() %>% dplyr::mutate(pathogen = pathogen_vars()),
           intervention_date = intervention_date()
         )
       }